Add Local Task Integrations
Problem
-------

As a developer I would like to include custom tasks to be used within a
single project. To do this I would need to write a minimum set of
configs and task classes and register them with dart_dev within the
dev.dart file. It would be nice if there was a local command that could
find all tasks within a tool director and allow dart_dev to run the
task.

Solution
--------

Provide the Local task within dart_dev that scans the the `tool`
directory of the current project and finds files that match the default
pattern of `[taskName]_task.[ext]`. Running `pub run dart_dev local
[taskName] [args]` will start a task with a defined executable in the
LocalConfig. Consumers can add additional executables that can handle a
given `ext` pattern.