CP-1835 Pass Arguments to LocalTasks
Problem
-------

Any arguments used when calling dart_dev are parsed by the dart dev
argument parser. It is not possible for locally discovered tasks to
register arguments preventing them from being parsed as valid arguments.
This prevents users from passing any arguments to the underlying local
task.

Solution
--------

Local tasks can now take a list of unparsed arguments as an optional
argument. If provided this list is passed to the invocation process
allowing the task to obtain a list of all parameters after the local
task name.

In addtion, when dart dev attempts to parse the arguments for the first
time, the list is scanned for the first commmand matching argument. It
is checked to confirm it is a local task. The list of arguments leading
up to the command are parsed allowing dart_dev global flags to be used
with local tasks.

Bonus
-----

A flag for expanding on symlinks within task search paths is provided in
the local configuration. This allows consumers to limit the search for
tasks to unlinked directories.