CPLAT-10490 - Crank down included files to whole directories that do not have excludes. 💼
This works for my use-case, but it could include whole directories that have linked files.  Is this going to be okay for the format tools?

CPLAT-10490 Allow an optional way to collapse directories. 🏋️‍♀️

This will allow existing behavior to continue for consumers, but will allow dpc to collapse directories to shorten argument list to over_react_format.  (over_react_format will be able to still expand any directories passed to it because of the default behavior.)

CPLAT-10490 Convert prints to finest logs, since they were necessary for me to be sure what the function was doing. 🧐

CPLAT-10490 - 🎯fmt

CPLAT-10490 🎯fmt again.

CPLAT-10490 Final cleanup ✨