Compile the dart_dev run script for better performance
After compilation, the startup time for subsequent `dart run dart_dev`
commands should be significantly improved.

This compiled exectuable is cached in `.dart_tool/dart_dev/` along with
a digest that is used to determine when recompilation is needed.

Packages with `tool/dart_dev/config.dart` files that import from their
own package's source files will be opted out of this compilation
optimization because there is no efficient way to track those imports
for the purpose of detecting the need to recompile.

This commit also adds a `clean` subcommand to `dart run dart_dev`
that will remove the files in `.dart_tool/dart_dev/`, in case a bug is
found in our compilation caching logic.